requirements.txt
PyQt6>=6.0.0
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pygame>=2.0.0
